Consolidate dialog logic Created some Composables to help create more custom dialogs in the future.
